changed exectionTime variable from int to long

+++ railo-java/railo-core/src/railo/runtime/PageContextImpl.java
@@ -195,7 +195,7 @@ public final class PageContextImpl extends PageContext implements Sizeable {
 	/**
 	 * Field <code>executionTime</code>
 	 */
-	protected int executionTime=0;
+	protected long executionTime=0;
 	
 	private HTTPServletRequestWrap req;
 	private HttpServletResponse rsp;
@@ -769,7 +769,7 @@ public final class PageContextImpl extends PageContext implements Sizeable {
 	public void doInclude(PageSource[] sources, boolean runOnce) throws PageException {
     	// debug
 		if(!gatewayContext && config.debug()) {
-			int currTime=executionTime;
+			long currTime=executionTime;
             long exeTime=0;
             long time=System.nanoTime();
             
@@ -778,7 +778,7 @@ public final class PageContextImpl extends PageContext implements Sizeable {
             DebugEntryTemplate debugEntry=debugger.getEntry(this,currentPage.getPageSource());
             try {
                 addPageSource(currentPage.getPageSource(),true);
-                debugEntry.updateFileLoadTime((int)(System.nanoTime()-time));
+                debugEntry.updateFileLoadTime((System.nanoTime()-time));
                 exeTime=System.nanoTime();
 
                 currentPage.call(this);
@@ -798,8 +798,8 @@ public final class PageContextImpl extends PageContext implements Sizeable {
 			}
 			finally {
 				includeOnce.add(currentPage.getPageSource());
-				int diff= ((int)(System.nanoTime()-exeTime)-(executionTime-currTime));
-			    executionTime+=(int)(System.nanoTime()-time);
+				long diff= ((System.nanoTime()-exeTime)-(executionTime-currTime));
+			    executionTime+=(System.nanoTime()-time);
 				debugEntry.updateExeTime(diff);
 				removeLastPageSource(true);
 			}	
@@ -2684,9 +2684,10 @@ public final class PageContextImpl extends PageContext implements Sizeable {
 		this.thread=thread;
 	}
 
+	// FUTURE add as long
     @Override
     public int getExecutionTime() {
-        return executionTime;
+        return (int)executionTime;
     }
 
     @Override
